tung

+++ routes/pizzaRoutes.js
@@ -1,5 +1,6 @@
 const express = require("express");
 const multer = require("multer");
+const { authorize, authenticateJWT } = require("../middleware/auth");
 const {
   createPizza,
   getPizzas,
@@ -9,7 +10,7 @@ const {
 } = require("../controllers/pizzaController");
 
 const router = express.Router();
-
+//Su ly file anh
 const storage = multer.diskStorage({
   destination: (req, file, cb) => {
     cb(null, "uploads/");
@@ -21,16 +22,23 @@ const storage = multer.diskStorage({
 
 const upload = multer({ storage: storage });
 
-router.post("/pizzas", upload.single("imgPiza"), async (req, res) => {
-  try {
-    const pizza = { ...req.body, imgPiza: req.file.path };
-    const newPizza = await createPizza(pizza);
-    res.status(201).json(newPizza);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
+//Tao piza
+router.post(
+  "/pizzas",
+  upload.single("imgPiza"),
+  authenticateJWT,
+  authorize(["A"]),
+  async (req, res) => {
+    try {
+      const pizza = { ...req.body, imgPiza: req.file.path };
+      const newPizza = await createPizza(pizza);
+      res.status(201).json(newPizza);
+    } catch (err) {
+      res.status(500).json({ error: err.message });
+    }
   }
-});
-
+);
+//Lay tat ca sp
 router.get("/pizzas", async (req, res) => {
   try {
     const pizzas = await getPizzas();
@@ -39,7 +47,7 @@ router.get("/pizzas", async (req, res) => {
     res.status(500).json({ error: err.message });
   }
 });
-
+// Chi tiet sp
 router.get("/pizzas/:idPiza", async (req, res) => {
   try {
     const pizza = await getPizzaById(req.params.idPiza);
@@ -48,27 +56,38 @@ router.get("/pizzas/:idPiza", async (req, res) => {
     res.status(500).json({ error: err.message });
   }
 });
-
-router.put("/pizzas/:idPiza", upload.single("imgPiza"), async (req, res) => {
-  try {
-    const pizza = {
-      ...req.body,
-      imgPiza: req.file ? req.file.path : undefined,
-    };
-    const updatedPizza = await updatePizza(req.params.idPiza, pizza);
-    res.status(200).json(updatedPizza);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
+// Cap nhat sp
+router.put(
+  "/pizzas/:idPiza",
+  upload.single("imgPiza"),
+  authenticateJWT,
+  authorize(["A"]),
+  async (req, res) => {
+    try {
+      const pizza = {
+        ...req.body,
+        imgPiza: req.file ? req.file.path : undefined,
+      };
+      const updatedPizza = await updatePizza(req.params.idPiza, pizza);
+      res.status(200).json(updatedPizza);
+    } catch (err) {
+      res.status(500).json({ error: err.message });
+    }
   }
-});
+);
 
-router.delete("/pizzas/:idPiza", async (req, res) => {
-  try {
-    const pizza = await deletePizza(req.params.idPiza);
-    res.status(200).json(pizza);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
+router.delete(
+  "/pizzas/:idPiza",
+  authenticateJWT,
+  authorize(["A"]),
+  async (req, res) => {
+    try {
+      const pizza = await deletePizza(req.params.idPiza);
+      res.status(200).json(pizza);
+    } catch (err) {
+      res.status(500).json({ error: err.message });
+    }
   }
-});
+);
 
 module.exports = router;
